+++ backend/hypatia/collect.py
@@ -3,13 +3,17 @@ Base class for tables data that use star names as their primary key (unique iden
 """
 import time
 
-import pymongo
+from pymongo import MongoClient
+from pymongo.cursor import Cursor
+from pymongo.errors import ServerSelectionTimeoutError
+from pymongo.results import DeleteResult, InsertOneResult, InsertManyResult, UpdateResult
+
 
 from hypatia.config import connection_string
 
 
 class BaseCollection:
-    client = pymongo.MongoClient(connection_string)
+    client = MongoClient(connection_string)
     validator = {
         '$jsonSchema': {
             'bsonType': 'object',
@@ -49,7 +53,7 @@ class BaseCollection:
         while count < tries:
             try:
                 self.server_info = self.client.server_info()
-            except pymongo.errors.ServerSelectionTimeoutError:
+            except ServerSelectionTimeoutError:
                 count += 1
                 time.sleep(1)
             else:
@@ -88,17 +92,16 @@ class BaseCollection:
     def collection_compound_index(self, index_dict: dict[str, int], unique: bool = False):
         self.collection.create_index(list(index_dict.items()), unique=unique)
 
-    def add_one(self, doc: dict | list | float | str | int) -> pymongo.results.InsertOneResult:
+    def add_one(self, doc: dict | list | float | str | int) -> InsertOneResult:
         return self.collection.insert_one(doc)
 
-    def add_many(self, docs: list[dict | list | float | str | int] | pymongo.cursor.Cursor) \
-            -> pymongo.results.InsertManyResult:
+    def add_many(self, docs: list[dict | list | float | str | int] | Cursor) -> InsertManyResult:
         return self.collection.insert_many(docs)
 
     def find_one(self, query: dict) -> dict:
         return self.collection.find_one(query)
 
-    def find_all(self, query: dict = None) -> pymongo.cursor.Cursor:
+    def find_all(self, query: dict = None) -> Cursor:
         if query is None:
             return self.collection.find()
         else:
@@ -107,7 +110,7 @@ class BaseCollection:
     def find_by_id(self, find_id: str) -> dict:
         return self.collection.find_one({'_id': find_id})
 
-    def remove_by_id(self, remove_id: str) -> pymongo.results.DeleteResult:
+    def remove_by_id(self, remove_id: str) -> DeleteResult:
         return self.collection.delete_one({'_id': remove_id})
 
 
@@ -119,5 +122,5 @@ class BaseStarCollection(BaseCollection):
     def create_indexes(self):
         self.collection_add_index(self.name_col, unique=True)
 
-    def update_timestamp(self, update_id: str) -> pymongo.results.UpdateResult:
+    def update_timestamp(self, update_id: str) -> UpdateResult:
         return self.collection.update({'_id': update_id}, {'$set': {'timestamp': time.time()}})
